backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
pandas==2.2.0
PyPDF2==3.0.1
pdfplumber==0.11.0
tabula-py==2.9.0
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.6.0
